+++ test/auto_parallel/CMakeLists.txt
@@ -127,10 +127,6 @@ if(WITH_DISTRIBUTE AND WITH_GPU)
                   test_semi_auto_parallel_hybrid_strategy)
   set_tests_properties(test_semi_auto_parallel_hybrid_strategy
                        PROPERTIES LABELS "RUN_TYPE=EXCLUSIVE" TIMEOUT 120)
-  py_test_modules(test_semi_auto_parallel_recompute MODULES
-                  test_semi_auto_parallel_recompute)
-  set_tests_properties(test_semi_auto_parallel_recompute
-                       PROPERTIES LABELS "RUN_TYPE=EXCLUSIVE" TIMEOUT 120)
   py_test_modules(test_gpt_with_newir MODULES test_gpt_with_newir)
   set_tests_properties(test_gpt_with_newir
                        PROPERTIES LABELS "RUN_TYPE=EXCLUSIVE" TIMEOUT 100)
