+++ test/auto_parallel/test_semi_auto_parallel_single_strategy.py
@@ -50,6 +50,16 @@ class TestSemiAutoParallelSingleStrategy(test_base.CommunicationTestDistBase):
                 user_defined_envs=envs,
             )
 
+    def test_simple_net_recompute(self):
+        envs_list = test_base.gen_product_envs_list(
+            self._default_envs, self._changeable_envs
+        )
+        for envs in envs_list:
+            self.run_test_case(
+                "semi_auto_parallel_recompute.py",
+                user_defined_envs=envs,
+            )
+
 
 if __name__ == "__main__":
     unittest.main()
